+++ packages/components/src/Icon/Icon.tsx
@@ -27,18 +27,21 @@
 import {
   color,
   CompatibleHTMLProps,
-  omitStyledProps,
   SizeXXSmall,
   SizeXSmall,
   SizeSmall,
   SizeMedium,
   SizeLarge,
+  omitStyledProps,
 } from '@looker/design-tokens'
-import React, { forwardRef, Ref } from 'react'
+import React, { forwardRef, Ref, ReactNode } from 'react'
 import styled from 'styled-components'
 /* eslint import/namespace: ['error', { allowComputed: true }] */
 import { Glyphs, IconNames } from '@looker/icons'
-import { simpleLayoutCSS, SimpleLayoutProps } from '../Layout/utils/simple'
+import {
+  sizeSimpleLayoutCSS,
+  SizeSimpleLayoutProps,
+} from '../Layout/utils/simple'
 
 export type IconSize =
   | SizeXXSmall
@@ -49,37 +52,49 @@ export type IconSize =
 
 export interface IconProps
   extends Omit<CompatibleHTMLProps<HTMLDivElement>, 'onClick'>,
-    SimpleLayoutProps {
+    SizeSimpleLayoutProps {
+  /**
+   * Display an icon/logo that is not available on our components list. Use artwork prop with an svg instead of Icon name.
+   */
+  artwork?: ReactNode
   color?: string
-  name: IconNames
+  name?: IconNames
 }
 
 export type { IconNames }
 
-const IconFactory = forwardRef(
-  ({ className, name, ...props }: IconProps, ref: Ref<HTMLDivElement>) => {
-    const Glyph = Glyphs[name]
-
+const IconLayout = forwardRef(
+  (
+    { artwork = undefined, name, ...props }: IconProps,
+    ref: Ref<HTMLDivElement>
+  ) => {
+    if ((artwork && name) || (!artwork && !name)) {
+      // eslint-disable-next-line no-console
+      console.warn('You may only specify name OR artwork, not both.')
+    }
+    const Glyph = name ? Glyphs[name] : 'div'
+    const value = artwork || (
+      <Glyph width="100%" height="100%" fill="currentColor" />
+    )
     return (
-      <div className={className} ref={ref} {...omitStyledProps(props)}>
-        <Glyph width="100%" height="100%" fill="currentColor" />
+      <div ref={ref} {...omitStyledProps(props)}>
+        {value}
       </div>
     )
   }
 )
 
-IconFactory.displayName = 'IconFactory'
+IconLayout.displayName = 'IconLayout'
 
-export const Icon = styled(IconFactory).attrs(({ size, height, width }) => ({
-  size: !height && !width ? size || 'medium' : undefined,
-}))`
-  ${simpleLayoutCSS}
+export const Icon = styled(IconLayout)<IconProps>`
+  ${sizeSimpleLayoutCSS}
   ${color}
-
   align-items: center;
-  display: inline-flex;
-  height: ${({ height }) => height};
-  width: ${({ width }) => width};
+
+  svg {
+    height: 100%;
+    width: 100%;
+  }
 `
 
-// Icon.defaultProps = { size: 'medium' }
+Icon.defaultProps = { display: 'inline-flex', size: 'medium' }