+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Tree`, `AccordionContent` no longer add automatic padding at the container level
   - Padding happens at the individual item level
+- `Icon` can now only be sized via `size` (no height/width since it's a square)
 
 ### Fixed
 
@@ -52,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Icon accepts artwork as props.
 - `Fieldset` supports an accordion mode via an `accordion` prop
 
 ### Changed